+++ src/lib/output/themes/MarkedPlugin.tsx
@@ -36,6 +36,9 @@ export class MarkedPlugin extends ContextAwareRendererComponent {
     @Option("markdownItOptions")
     accessor markdownItOptions!: Record<string, unknown>;
 
+    @Option("markdownLinkExternal")
+    accessor markdownLinkExternal!: boolean;
+
     private parser?: markdown;
 
     /**
@@ -267,6 +270,17 @@ export class MarkedPlugin extends ContextAwareRendererComponent {
             const token = tokens[idx];
             const href = token.attrGet("href")?.replace(/^#(?:md:)?(.+)/, "#md:$1");
             if (href) {
+                // Note: This doesn't catch @link tags to reflections as those
+                // will be relative links. This will likely have to change with
+                // the introduction of support for customized routers whenever
+                // that becomes a real thing.
+                if (this.markdownLinkExternal && /https?:\/\//i.test(href)) {
+                    token.attrSet("target", "_blank");
+                    const classes = token.attrGet("class")?.split(" ") || [];
+                    classes.push("external");
+                    token.attrSet("class", classes.join(" "));
+                }
+
                 token.attrSet("href", href);
             }
             return self.renderToken(tokens, idx, options);