+++ src/test/converter2/issues/gh2681.ts
@@ -1,0 +1,4 @@
+/**
+ * {@link Generator}
+ */
+export const bug = 123;